backend/pyproject.toml
[project]
name = "swarmops-backend"
version = "0.1.0"
description = "SwarmOps multi-agent fintech commentary backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.0",
    "langchain-aws>=1.3.1",
    "langchain-core>=1.2.16",
    "langgraph>=1.0.10",
    "pydantic-settings>=2.13.1",
    "sse-starlette>=3.3.2",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
